refactored samples test
